Disable testing of README.rst with py2.x.
As it is, README.rst doctests fail with Python2.x as they do not
parse unicode chars correctly. There is no easy way to declare an
encoding for doctests .rst files (not a problem in docstrings of
modules).

We _could_ make examples in README.rst runnable under python2 _and_
python3. But the resulting code would be nearly unreadable and
could stop readers from getting quickly what the point is.

The README is still tested with all python 3.x flavours.